LLM-rag-backend/phi3-rag-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
llama-cpp-python==0.2.87
sentence-transformers==3.0.1
faiss-cpu>=1.13.0,<1.14.0
pymupdf==1.24.11
python-multipart==0.0.9
